requirements.txt
fastapi==0.65.2
fastapi-utils==0.2.1
imageio==2.19.5
Pillow==9.0.1
python-dateutil==2.8.1
pytz==2021.1
uvicorn==0.11.7
requests==2.25.1
requests-cache==0.9.3
